API/app/requirements.txt
fastapi==0.62.0
Pillow==8.0.1
pydantic==1.7.3
pymongo==3.11.2
starlette==0.13.6
treepoem==3.7.0
